# Conflicts:
#	core/src/joker_effect_processor.rs
#	core/src/static_joker.rs

+++ core/src/scaling_joker_custom.rs
@@ -1,9 +1,9 @@
-use crate::scaling_joker::{ScalingJoker, ScalingEvent, ScalingEffectType};
-use crate::joker::{Joker, JokerId, JokerRarity, JokerEffect, GameContext};
-use crate::joker_state::JokerState;
-use crate::hand::SelectHand;
 use crate::card::Card;
+use crate::hand::SelectHand;
+use crate::joker::{GameContext, Joker, JokerEffect, JokerId, JokerRarity};
+use crate::joker_state::JokerState;
 use crate::rank::HandRank;
+use crate::scaling_joker::{ScalingEffectType, ScalingEvent, ScalingJoker};
 
 /// Custom implementations for scaling jokers that require special logic
 /// beyond the basic ScalingJoker framework
@@ -41,9 +41,11 @@ impl Joker for GreenJoker {
 
     fn on_hand_played(&self, context: &mut GameContext, hand: &SelectHand) -> JokerEffect {
         // Increment value for any hand played
-        context.joker_state_manager.update_state(self.id(), |state| {
-            state.accumulated_value += 1.0;
-        });
+        context
+            .joker_state_manager
+            .update_state(self.id(), |state| {
+                state.accumulated_value += 1.0;
+            });
 
         // Return current effect
         let current_value = context
@@ -55,9 +57,11 @@ impl Joker for GreenJoker {
 
     fn on_discard(&self, context: &mut GameContext, cards: &[Card]) -> JokerEffect {
         // Decrease value for each discard action (not per card)
-        context.joker_state_manager.update_state(self.id(), |state| {
-            state.accumulated_value = (state.accumulated_value - 1.0).max(0.0);
-        });
+        context
+            .joker_state_manager
+            .update_state(self.id(), |state| {
+                state.accumulated_value = (state.accumulated_value - 1.0).max(0.0);
+            });
 
         JokerEffect::new()
     }
@@ -101,9 +105,11 @@ impl Joker for SquareJoker {
     fn on_hand_played(&self, context: &mut GameContext, hand: &SelectHand) -> JokerEffect {
         // Check if hand has exactly 4 cards
         if hand.len() == 4 {
-            context.joker_state_manager.update_state(self.id(), |state| {
-                state.accumulated_value += 4.0;
-            });
+            context
+                .joker_state_manager
+                .update_state(self.id(), |state| {
+                    state.accumulated_value += 4.0;
+                });
         }
 
         // Return current effect
@@ -152,7 +158,7 @@ impl Joker for BullJoker {
 
     fn on_hand_played(&self, context: &mut GameContext, _hand: &SelectHand) -> JokerEffect {
         // Effect is based on current money, not accumulated
-        let chips_bonus = context.money * 2;
+        let chips_bonus = (context.money as f64 * 2.0) as i32;
         JokerEffect::new().with_chips(chips_bonus)
     }
 
@@ -194,7 +200,7 @@ impl Joker for BootstrapsJoker {
 
     fn on_hand_played(&self, context: &mut GameContext, _hand: &SelectHand) -> JokerEffect {
         // Effect is based on current money divided by 5
-        let mult_bonus = (context.money / 5) * 2;
+        let mult_bonus = ((context.money as f64 / 5.0) * 2.0) as i32;
         JokerEffect::new().with_mult(mult_bonus)
     }
 
@@ -238,9 +244,9 @@ impl Joker for BannerJoker {
         // Calculate remaining discards (assuming 3 base discards per round)
         let base_discards: usize = 3; // This should come from game config
         let remaining_discards = base_discards.saturating_sub(context.discards_used as usize);
-        let chips_bonus = remaining_discards * 30;
+        let chips_bonus = (remaining_discards * 30) as i32;
         
-        JokerEffect::new().with_chips(chips_bonus as i32)
+        JokerEffect::new().with_chips(chips_bonus)
     }
 
     fn initialize_state(&self, _context: &GameContext) -> JokerState {
@@ -281,12 +287,14 @@ impl Joker for CeremonialDagger {
 
     fn on_blind_start(&self, context: &mut GameContext) -> JokerEffect {
         // Double the mult multiplier when blind starts
-        context.joker_state_manager.update_state(self.id(), |state| {
-            state.accumulated_value *= 2.0;
-            if state.accumulated_value == 0.0 {
-                state.accumulated_value = 2.0; // Start at 2x if was 0
-            }
-        });
+        context
+            .joker_state_manager
+            .update_state(self.id(), |state| {
+                state.accumulated_value *= 2.0;
+                if state.accumulated_value == 0.0 {
+                    state.accumulated_value = 2.0; // Start at 2x if was 0
+                }
+            });
 
         JokerEffect::new()
     }
@@ -302,9 +310,11 @@ impl Joker for CeremonialDagger {
 
     fn on_round_end(&self, context: &mut GameContext) -> JokerEffect {
         // Reset to base value at round end
-        context.joker_state_manager.update_state(self.id(), |state| {
-            state.accumulated_value = 1.0;
-        });
+        context
+            .joker_state_manager
+            .update_state(self.id(), |state| {
+                state.accumulated_value = 1.0;
+            });
 
         JokerEffect::new()
     }
@@ -351,15 +361,18 @@ impl Joker for MysticSummit {
             Ok(made_hand) => made_hand.rank,
             Err(_) => HandRank::HighCard, // Fallback to high card if evaluation fails
         };
-        
+
         // Check if this hand type was played before
         let played_count = context.get_hand_type_count(hand_rank);
-        
+
         // If this is the first time playing this hand type, increment
-        if played_count == 1 { // Count includes current hand, so 1 means first time
-            context.joker_state_manager.update_state(self.id(), |state| {
-                state.accumulated_value += 15.0;
-            });
+        if played_count == 1 {
+            // Count includes current hand, so 1 means first time
+            context
+                .joker_state_manager
+                .update_state(self.id(), |state| {
+                    state.accumulated_value += 15.0;
+                });
         }
 
         // Return current effect
@@ -406,8 +419,8 @@ pub fn get_custom_scaling_joker_by_id(id: JokerId) -> Option<Box<dyn Joker>> {
 mod tests {
     use super::*;
     use crate::joker_state::JokerStateManager;
-    use std::sync::Arc;
     use std::collections::HashMap;
+    use std::sync::Arc;
 
     fn create_test_context() -> GameContext<'static> {
         // Placeholder - in real tests we'd need proper context setup
@@ -421,15 +434,23 @@ mod tests {
         assert_eq!(joker.name(), "Green Joker");
     }
 
-    #[test] 
+    #[test]
     fn test_all_custom_jokers_created() {
         let jokers = create_all_custom_scaling_jokers();
-        assert_eq!(jokers.len(), 7, "Should create exactly 7 custom scaling jokers");
-        
+        assert_eq!(
+            jokers.len(),
+            7,
+            "Should create exactly 7 custom scaling jokers"
+        );
+
         // Test that all jokers have unique IDs
         let mut ids = std::collections::HashSet::new();
         for joker in &jokers {
-            assert!(ids.insert(joker.id()), "Duplicate joker ID found: {:?}", joker.id());
+            assert!(
+                ids.insert(joker.id()),
+                "Duplicate joker ID found: {:?}",
+                joker.id()
+            );
         }
     }
